% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{citation.rnaseqExplorer}
\alias{citation.rnaseqExplorer}
\title{Citation for rnaseqExplorer}
\usage{
citation.rnaseqExplorer()
}
\description{
Provides the citation information for the rnaseqExplorer package.
}
