% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{get_gene_details}
\alias{get_gene_details}
\title{getting gene annotation}
\usage{
get_gene_details(ensembl_id)
}
\arguments{
\item{ensembl_id}{}
}
\value{
gene annotation details
}
\description{
exporting gene annotatioin based on ensembl id
}
\examples{
temp1 <- get_gene_details('ENSG00000139618')
}
