% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchApp}
\alias{launchApp}
\title{Launch Shiny Application}
\usage{
launchApp()
}
\arguments{
\item{ui}{The UI layout of the Shiny application}

\item{server}{The server logic of the Shiny application}
}
\description{
This function initializes and runs the Shiny application by combining the UI and server components.
It serves as a wrapper around the `shinyApp()` function.
}
\examples{
launchApp()
}
