% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\docType{data}
\name{required_packages}
\alias{required_packages}
\title{Launch Shiny Application}
\format{
An object of class \code{character} of length 10.
}
\usage{
required_packages
}
\arguments{
\item{ui}{The UI layout of the Shiny application}

\item{server}{The server logic of the Shiny application}
}
\description{
This function initializes and runs the Shiny application by combining the UI and server components.
It serves as a wrapper around the `shinyApp()` function.
}
\examples{
launchApp()
}
\keyword{datasets}
