% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Launch Shiny Application}
\usage{
run_app(ui, server)
}
\arguments{
\item{ui}{The UI layout of the Shiny application}

\item{server}{The server logic of the Shiny application}
}
\description{
This function initializes and runs the Shiny application by combining the UI and server components.
It serves as a wrapper around the `shinyApp()` function.
}
\examples{
run_my_app()
}
