% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{Defining server logic for handling user interactions and data}
\usage{
server(input, output, session)
}
\description{
Implements the server-side logic, including reactive expressions and observers,
to process and display user information in the Shiny app.
}
\examples{
server <- function(input, output, session) {
  server_logic()
}
}
