% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ui}
\alias{ui}
\title{Creating the Shiny UI for displaying user information}
\usage{
ui()
}
\description{
Generates a Shiny UI layout for displaying user details such as name, email, and activities.
This function defines the structure and layout of the app's UI.
}
\examples{
ui <- create_user_ui()
}
