% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch Shiny Application}
\usage{
run_app()
}
\description{
This function initializes and runs the Shiny application by combining the UI and server components.
It serves as a wrapper around the `shinyApp()` function.
}
\examples{
run_app()
}
