% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{ui}
\alias{ui}
\title{Creating the Shiny UI for displaying user information}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 5.
}
\usage{
ui
}
\description{
Generates a Shiny UI layout for displaying user details such as name, email, and activities.
This function defines the structure and layout of the app's UI.
}
\examples{
ui <- create_user_ui()
}
\keyword{datasets}
